feat: fix line chart stats flashing + add generalize plugins
