+++ src/components/charts/LineChart.tsx
@@ -13,6 +13,7 @@ import {
 } from "chart.js";
 import React, { useCallback, useEffect, useMemo, useRef } from "react";
 import { ReactChart } from "../ReactChart";
+import { LineChartHorizontalReferenceLine, plugins } from "./chartHelpers";
 
 Chart.register(LineController, LineElement, LinearScale, LogarithmicScale, CategoryScale, PointElement, Filler);
 
@@ -48,12 +49,7 @@ export interface LineChartProps {
   onMouseOver?: (index: number) => void;
   activeIndex?: number;
   useLogarithmicScale?: boolean;
-  horizontalReferenceLines?: {
-    value: number;
-    color: string;
-    dash?: number[];
-    label?: string;
-  }[];
+  horizontalReferenceLines?: LineChartHorizontalReferenceLine[];
   // Props for custom y-axis range
   yAxisMin?: number;
   yAxisMax?: number;
@@ -155,264 +151,37 @@ const LineChart = React.memo(
       [data, makeLineGradients, makeAreaGradients, xKey],
     );
 
-    const gradientPlugin = useMemo(() => {
-      return {
-        id: "customGradientShift",
-        beforeUpdate: (chart) => {
-          const ctx = chart.ctx;
-          const activeIndex = activeIndexRef.current;
-          if (ctx && typeof activeIndex === "number") {
-            const grayColor = "rgba(128, 128, 128, 0.1)"; // Define your gray color here
+    // ---------- PLUGINS ----------
 
-            for (let i = 0; i < chart.data.datasets.length; ++i) {
-              const dataset = chart.data.datasets[i];
-              const lineGradient = makeLineGradients[i](ctx, 1);
-              const areaGradient = makeAreaGradients ? makeAreaGradients[i](ctx, 1) : null;
-
-              // Apply gradients to the entire dataset
-              dataset.borderColor = lineGradient;
-              dataset.backgroundColor = areaGradient;
-
-              // Use segment configuration for conditional coloring
-              dataset.segment = {
-                borderColor: (ctx) => {
-                  const dataIndex = ctx.p0DataIndex;
-                  const isAfterActiveIndex = dataIndex >= activeIndex;
-                  return isAfterActiveIndex ? grayColor : undefined;
-                },
-                backgroundColor: (ctx) => {
-                  const dataIndex = ctx.p0DataIndex;
-                  const isAfterActiveIndex = dataIndex >= activeIndex;
-                  return isAfterActiveIndex ? grayColor : undefined;
-                },
-              };
-            }
-          }
-        },
-      };
-    }, [makeLineGradients, makeAreaGradients]); // Removed morningIndex from dependencies
+    const gradientPlugin = useMemo(
+      () => plugins.gradientShift(activeIndexRef, makeLineGradients, makeAreaGradients),
+      [makeLineGradients, makeAreaGradients],
+    ); // Removed morningIndex from dependencies
 
     const fillArea = !!makeAreaGradients && !!makeAreaGradients.length;
 
-    // const referenceDotPlugin: Plugin = useMemo<Plugin>(() => {}, []);
+    const verticalLinePlugin: Plugin = useMemo(() => plugins.verticalLine(activeIndexRef, fillArea), [fillArea]);
 
-    const verticalLinePlugin: Plugin = useMemo<Plugin>(
-      () => ({
-        id: "customVerticalLine",
-        afterDraw: (chart: Chart) => {
-          const ctx = chart.ctx;
-          const activeIndex = activeIndexRef.current;
-          if (ctx) {
-            ctx.save();
-            ctx.setLineDash([4, 4]);
+    const horizontalReferenceLinePlugin: Plugin = useMemo(() => {
+      return plugins.horizontalReferenceLine(activeIndexRef, horizontalReferenceLines);
+    }, [horizontalReferenceLines]);
 
-            // Draw the vertical line at morningIndex
-            if (typeof activeIndex === "number") {
-              const morningDataPoint = chart.getDatasetMeta(0).data[activeIndex];
-              if (morningDataPoint) {
-                const { x } = morningDataPoint.getProps(["x"], true);
-                ctx.beginPath();
-                ctx.moveTo(x, chart.chartArea.top);
-                ctx.lineTo(x, chart.chartArea.bottom);
-                ctx.strokeStyle = "#D9AD0F"; // Use a different color for the morning line
-                ctx.lineWidth = 1.5;
-                ctx.stroke();
-              }
-            }
+    const selectionPointPlugin = useMemo(() => {
+      return plugins.selectionPoint(activeIndexRef, fillArea);
+    }, [fillArea]);
 
-            // Draw the vertical line for the active element (hovered point)
-            const activeElements = chart.getActiveElements();
-            if (activeElements.length > 0) {
-              const activeElement = activeElements[0];
-              const datasetIndex = activeElement.datasetIndex;
-              const index = activeElement.index;
-              const dataPoint = chart.getDatasetMeta(datasetIndex).data[index];
+    // create a ref to the onMouseOver function
+    const onMouseOverRef = useRef<typeof onMouseOver>(onMouseOver);
 
-              if (dataPoint) {
-                const { x } = dataPoint.getProps(["x"], true);
-                ctx.beginPath();
-                ctx.moveTo(x, chart.chartArea.top);
-                ctx.lineTo(x, chart.chartArea.bottom);
-                ctx.strokeStyle = fillArea ? "#D9AD0F" : "#246645";
-                ctx.lineWidth = 1.5;
-                ctx.stroke();
-              }
-            }
+    // update the ref when the onMouseOver function changes
+    useEffect(() => {
+      onMouseOverRef.current = onMouseOver;
+    }, [onMouseOver]);
 
-            ctx.restore();
-          }
-        },
-      }),
-      [fillArea], // Removed morningIndex from dependencies
-    );
-
-    const horizontalReferenceLinePlugin: Plugin = useMemo<Plugin>(
-      () => ({
-        id: "horizontalReferenceLine",
-        afterDraw: (chart: Chart) => {
-          const ctx = chart.ctx;
-          if (!ctx || horizontalReferenceLines.length === 0) return;
-
-          ctx.save();
-
-          // Draw each horizontal reference line
-          horizontalReferenceLines.forEach((line) => {
-            const yScale = chart.scales.y;
-            const y = yScale.getPixelForValue(line.value);
-
-            // Only draw if within chart area
-            if (y >= chart.chartArea.top && y <= chart.chartArea.bottom) {
-              ctx.beginPath();
-              if (line.dash) {
-                ctx.setLineDash(line.dash);
-              } else {
-                ctx.setLineDash([4, 4]); // Default dash pattern
-              }
-              ctx.moveTo(chart.chartArea.left, y);
-              ctx.lineTo(chart.chartArea.right, y);
-              ctx.strokeStyle = line.color;
-              ctx.lineWidth = 1;
-              ctx.stroke();
-
-              // Reset dash pattern
-              ctx.setLineDash([]);
-
-              // Add label if provided
-              if (line.label) {
-                ctx.font = "12px Arial";
-                ctx.fillStyle = line.color;
-
-                // Measure text width to ensure it doesn't get cut off
-                const textWidth = ctx.measureText(line.label).width;
-                const rightPadding = 10; // Padding from right edge
-
-                // Position the label at the right side of the chart with padding
-                const labelX = chart.chartArea.right - textWidth - rightPadding;
-                const labelPadding = 5; // Padding between line and text
-                const textHeight = 12; // Approximate height of the text
-
-                // Check if the line is too close to the top of the chart
-                const isNearTop = y - textHeight - labelPadding < chart.chartArea.top;
-
-                // Check if the line is too close to the bottom of the chart
-                const isNearBottom = y + textHeight + labelPadding > chart.chartArea.bottom;
-
-                // Set text alignment
-                ctx.textAlign = "left";
-
-                // Position the label based on proximity to chart edges
-                // biome-ignore lint/suspicious/noExplicitAny:
-                let labelY: any;
-                ctx.textBaseline = "bottom";
-                labelY = y - labelPadding;
-                if (isNearTop) {
-                  ctx.textBaseline = "top";
-                  labelY = y + labelPadding;
-                } else if (isNearBottom) {
-                  labelY = y - labelPadding;
-                }
-                ctx.fillText(line.label, labelX, labelY);
-              }
-            }
-          });
-
-          ctx.restore();
-        },
-      }),
-      [horizontalReferenceLines],
-    );
-
-    const selectionPointPlugin: Plugin = useMemo<Plugin>(
-      () => ({
-        id: "customSelectPoint",
-        afterDraw: (chart: Chart) => {
-          const ctx = chart.ctx;
-          const activeIndex = activeIndexRef.current;
-          if (!ctx) return;
-
-          // Define the function to draw the selection point
-          const drawSelectionPoint = (x: number, y: number) => {
-            ctx.save();
-            ctx.fillStyle = fillArea ? "#D9AD0F" : "#246645";
-            ctx.strokeStyle = fillArea ? "#D9AD0F" : "#246645";
-            ctx.lineWidth = 1;
-
-            const rectWidth = 6;
-            const rectHeight = 6;
-            const cornerRadius = 4;
-
-            ctx.beginPath();
-            ctx.moveTo(x - rectWidth / 2 + cornerRadius, y - rectHeight / 2);
-            ctx.lineTo(x + rectWidth / 2 - cornerRadius, y - rectHeight / 2);
-            ctx.quadraticCurveTo(
-              x + rectWidth / 2,
-              y - rectHeight / 2,
-              x + rectWidth / 2,
-              y - rectHeight / 2 + cornerRadius,
-            );
-            ctx.lineTo(x + rectWidth / 2, y + rectHeight / 2 - cornerRadius);
-            ctx.quadraticCurveTo(
-              x + rectWidth / 2,
-              y + rectHeight / 2,
-              x + rectWidth / 2 - cornerRadius,
-              y + rectHeight / 2,
-            );
-            ctx.lineTo(x - rectWidth / 2 + cornerRadius, y + rectHeight / 2);
-            ctx.quadraticCurveTo(
-              x - rectWidth / 2,
-              y + rectHeight / 2,
-              x - rectWidth / 2,
-              y + rectHeight / 2 - cornerRadius,
-            );
-            ctx.lineTo(x - rectWidth / 2, y - rectHeight / 2 + cornerRadius);
-            ctx.quadraticCurveTo(
-              x - rectWidth / 2,
-              y - rectHeight / 2,
-              x - rectWidth / 2 + cornerRadius,
-              y - rectHeight / 2,
-            );
-            ctx.closePath();
-
-            ctx.fill();
-            ctx.stroke();
-            ctx.restore();
-          };
-
-          // Draw selection point for the hovered data point
-          const activeElements = chart.getActiveElements();
-          for (const activeElement of activeElements) {
-            const datasetIndex = activeElement.datasetIndex;
-            const index = activeElement.index;
-            const dataPoint = chart.getDatasetMeta(datasetIndex).data[index];
-
-            if (dataPoint) {
-              const { x, y } = dataPoint.getProps(["x", "y"], true);
-              drawSelectionPoint(x, y);
-            }
-          }
-
-          // Draw selection point for the morningIndex
-          if (typeof activeIndex === "number") {
-            const dataPoint = chart.getDatasetMeta(0).data[activeIndex];
-            if (dataPoint) {
-              const { x, y } = dataPoint.getProps(["x", "y"], true);
-              drawSelectionPoint(x, y);
-            }
-          }
-        },
-      }),
-      [fillArea], // Removed morningIndex from dependencies
-    );
-
-    const selectionCallbackPlugin: Plugin = useMemo<Plugin>(
-      () => ({
-        id: "selectionCallback",
-        afterDraw: (chart: Chart) => {
-          onMouseOver?.(chart.getActiveElements()[0]?.index);
-        },
-      }),
-      [],
-    );
+    // Plugin to handle mouse over
+    const selectionCallbackPlugin: Plugin = useMemo(() => {
+      return plugins.selectionCallback(onMouseOverRef.current);
+    }, []);
 
     const chartOptions: ChartOptions = useMemo(() => {
       return {
@@ -445,7 +214,7 @@ const LineChart = React.memo(
             grid: {
               display: true,
               color: (context) => {
-                const tickLabel = context.tick && context.tick.label;
+                const tickLabel = context.tick?.label;
                 if (typeof activeIndex === "number") {
                   if (tickLabel && tickLabel !== "") {
                     return "rgba(0, 0, 0, 0.1)";
